+++ TF_Settings_Web/src/Components/TopBar/TabBar/TabBar.module.scss
@@ -2,7 +2,7 @@
 
 .tab-bar {
     position: sticky;
-    top: v.$control-bar-height;
+    top: v.$header-height;
     display: flex;
     width: 100vw;
     height: 10vh;
