+++ TF_Settings_Web/src/Components/TopBar/VersionIndicator/VersionIndicator.module.scss
