+++ TF_Settings_Web/src/Components/TopBar/index.ts
@@ -1,7 +1,7 @@
-import ControlBar from './ControlBar';
+import Header from './Header/Header';
 import StatusIndicator from './StatusIndicator/StatusIndicator';
 import TabBar from './TabBar/TabBar';
 import TabSelector from './TabSelector/TabSelector';
 import VersionIndicator from './VersionIndicator/VersionIndicator';
 
-export { ControlBar, StatusIndicator, TabSelector, VersionIndicator, TabBar };
+export { Header, StatusIndicator, TabSelector, VersionIndicator, TabBar };
