+++ TF_Settings_Web/src/Pages/About/About.module.scss
@@ -21,7 +21,6 @@
 }
 
 .info-table {
-    $numCols: 1;
     background: v.$dark-grey-gradient;
     box-shadow: 0px 2px 10px #00000099;
     border: 1px solid #ffffff1a;
