+++ TF_Settings_Web/src/Pages/Camera/CameraSetupScreen.tsx
@@ -11,7 +11,7 @@ import { TrackingServiceState } from 'TouchFree/src/TouchFreeToolingTypes';
 import { QuickSetupIcon, ManualSetupIcon, CameraMaskingIcon } from '@/Images';
 
 import { Alert, DocsLink, HorizontalIconTextButton, VerticalIconTextButton } from '@/Components';
-import { TabBar } from '@/Components/ControlBar';
+import { TabBar } from '@/Components/TopBar';
 
 const classes = classnames.bind(styles);
