+++ TF_Settings_Web/src/variables.module.scss
@@ -8,7 +8,7 @@ $dark-grey: #363636;
 $dark-grey-gradient: linear-gradient(180deg, $dark-grey 0%, #2c2c2c 100%);
 $outline-button-gradient: linear-gradient(180deg, $ultraleap-green, #00cdcf);
 $base-font: MontserratRegular, Arial;
-$control-bar-height: 8vh;
+$header-height: 8vh;
 
 :export {
     appWidth: $app-width;
